backend/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
dj-rest-auth==2.2.1
Django==4.0
django-admin-interface==0.18.3
django-allauth==0.47.0
django-colorfield==0.6.0
django-cors-headers==3.10.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-phonenumber-field==6.0.0
djangorestframework==3.13.1
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
oauthlib==3.1.1
packaging==21.3
phonenumbers==8.12.39
Pillow==8.4.0
psycopg2-binary==2.9.2
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.6
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.19
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.7
django-extensions==3.1.5
drf-writable-nested==0.6.3
celery==5.2.3
sqlalchemy==1.4.31
django-celery-results==2.2.0
suds-py3==1.4.5.0
checkout-sdk==2.0b9
